refactor(builders): Rearrange packages.
